refactor: removed unused dep, optimized code

+++ src/main/java/com/bkr/shopen/controller/UserController.java
@@ -38,7 +38,7 @@ public class UserController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<User> getUserById(@PathVariable Integer id) {
+    public ResponseEntity<User> getUserById(@PathVariable Long id) {
         User user = userService.getUserById(id);
         
         return ResponseEntity.ok(user);
